Moved all source files into point_one/fusion_engine/ subdirectory/namespace.
