+++ CMakeLists.txt
@@ -23,7 +23,7 @@ include_directories(${PROJECT_SOURCE_DIR}/src)
 
 # All messages and supporting code.
 add_library(fusion_engine_client
-            src/point_one/messages/crc.cc)
+            src/point_one/fusion_engine/messages/crc.cc)
 
 # Install targets.
 install(TARGETS fusion_engine_client
