+++ src/point_one/fusion_engine/messages/core.h
@@ -4,5 +4,5 @@
 
 #pragma once
 
-#include "point_one/messages/defs.h"
-#include "point_one/messages/solution.h"
+#include "point_one/fusion_engine/messages/defs.h"
+#include "point_one/fusion_engine/messages/solution.h"
